refactor: Replace deprecated fastutil Set.rem with Set.remove
On fastutil's *Set interfaces, rem(primitive) is deprecated in favor of
remove(primitive). Update the SetInclusionKernel and SSM hand-written
Char/Float sources, the UpdatableTable / TestAggBy / IntrusiveSoftLRU
call sites, and re-run the SetInclusionKernel and SegmentedSortedMultiset
replicators. The SSM replicator's Object post-processing no longer needs
the .rem(->.remove( substitution since the Char source itself now uses
remove. SortedRangesTest keeps vs.rem(long) because vs is a LongArrayList
and List.remove(int) removes by index, not by value — LongCollection.rem
is not deprecated.

+++ engine/table/src/main/java/io/deephaven/engine/table/impl/select/analyzers/SelectAndViewAnalyzer.java
@@ -199,7 +199,7 @@ public class SelectAndViewAnalyzer implements LogOutputAppendable {
                         throw new IllegalArgumentException(
                                 "Respected barrier, " + barrier + ", is not defined for " + sc.getName());
                     }
-                    execDeps.add(layerForBarrier);
+                    execDeps.add(layerForBarrier.intValue());
                 }
             }
